small change to allow stop <- false
This will allow for a little select{} magic to check if something is still playing without killing the current playback:
```
select{
case stop <- false:
   // still alive and didn't kill
case stop <- true:
   // was alive, but just killed
default:
   // was already dead
}
```

+++ dgvoice.go
@@ -160,8 +160,12 @@ func PlayAudioFile(v *discordgo.VoiceConnection, filename string, stop <-chan bo
 	}
 
 	go func() {
-		<-stop
-		err = run.Process.Kill()
+		for {
+			kill := <-stop
+			if kill {
+				err = run.Process.Kill()
+			}
+		}
 	}()
 
 	// Send "speaking" packet over the voice websocket
